pyproject.toml
[project]
name = "eeg-be"
version = "0.1.0"
description = "EEG backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "minio>=7.2,<8.0",
    "python-multipart>=0.0.6,<0.1.0",
    "PyJWT>=2.8,<3.0",
    "email-validator>=2.1,<3.0"
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
namespaces = false
